+++ ChangeLog.md
@@ -5,3 +5,6 @@
 - Initial release
 
 ## Unreleased changes
+
+- assertion over infinitive data structure or not terminating predicates do not hang all test and timeout is reported
+- the timeout for assertion is configurable now